Change return value of Shortest Path function to tuple

+++ Graph.cpp
@@ -20,16 +20,16 @@ Graph::Graph(size_t num_verticies) {
 	}
 }
 
-Graph::Graph(size_t num_verticies, double density, uint32_t seed) : Graph(num_verticies){
+Graph::Graph(size_t num_verticies, double density, uint64_t seed) : Graph(num_verticies){
     default_random_engine generator(seed);
     uniform_int_distribution<size_t> nodes_distribution(0, num_verticies_ - 1);
-    uniform_int_distribution<uint32_t> weights_distribution(1, num_verticies_);
+    uniform_int_distribution<uint64_t> weights_distribution(1, num_verticies_);
 
     size_t num_adjacent = static_cast<size_t>(density * static_cast<double>(num_verticies_));
 
     for(vector<Node>::iterator it = V_.begin(); it != V_.end(); ++it){
         it->adjacent_nodes_ = vector<size_t>(num_adjacent);
-        it->adjacent_weights_ = vector<uint32_t>(num_adjacent);
+        it->adjacent_weights_ = vector<uint64_t>(num_adjacent);
 
         for(size_t i = 0; i < num_adjacent; i++){
         	it->adjacent_nodes_[i] = nodes_distribution(generator);
